requirements.txt
torch 
torchvision 
transformers 
pillow 
tqdm 
timm
tensorflow
numpy
streamlit